fix(tui): widen TUI widget token fields u32→u64
All token-related fields in TUI widgets widened to match
the runtime's u64 token pipeline (rig-core returns u64):

- InferStreamData: tokens_in, tokens_out, max_tokens
- InferBox: tokens_in, tokens_out, thinking_tokens
- AgentBox: tokens_in, tokens_out
- TokenUsage: input_tokens, output_tokens, cache tokens
- StreamingProgress: tokens_generated, tokens_expected
- ModelInfo: max_tokens
- ActivityItem: tokens tuple
- ActiveOperation: tokens_in, tokens_out
- ChatView: start_infer_stream, append_infer_content
- format_number, calculate_cost, format_tokens signatures

+++ tools/nika/src/tui/views/chat/task_boxes.rs
@@ -233,7 +233,7 @@ impl ChatView {
     /// Don't create InferStream box - use streaming_decrypt for visual effect instead.
     /// The streaming decrypt provides the matrix reveal effect, while InferStream boxes
     /// were redundant and blocked the decrypt from showing.
-    pub fn start_infer_stream(&mut self, model: &str, _tokens_in: u32, _max_tokens: u32) {
+    pub fn start_infer_stream(&mut self, model: &str, _tokens_in: u64, _max_tokens: u64) {
         // Don't add to inline_content - let streaming_decrypt handle the visual
         // The matrix decrypt effect is the WOW feature, not the INFER boxes
 
@@ -246,7 +246,7 @@ impl ChatView {
 
     /// Append content to current inference stream
     /// Updated to handle TaskBox::Infer (tokens only) - Matrix effect handles text
-    pub fn append_infer_content(&mut self, chunk: &str, tokens_out: u32) {
+    pub fn append_infer_content(&mut self, chunk: &str, tokens_out: u64) {
         // Update TaskBox::Infer token count (if present)
         // Handle Task(TaskBox::Infer)
         for content in self.inline_content.iter_mut().rev() {
